chore: fix changes after code review
