+++ frontend/src/features/public-form/components/FormEndPage/FormEndPageContainer.tsx
@@ -71,17 +71,17 @@ export const FormEndPageContainer = (): JSX.Element | null => {
         colorTheme={form.startPage.colorTheme}
       />
     )
-  } else
-    return (
-      <Box py={{ base: '1.5rem', md: '2.5rem' }} w="100%">
-        <FormEndPage
-          colorTheme={form.startPage.colorTheme}
-          submissionData={submissionData}
-          formTitle={form.title}
-          endPage={form.endPage}
-          isFeedbackSubmitted={isFeedbackSubmitted}
-          handleSubmitFeedback={handleSubmitFeedback}
-        />
-      </Box>
-    )
+  }
+  return (
+    <Box py={{ base: '1.5rem', md: '2.5rem' }} w="100%">
+      <FormEndPage
+        colorTheme={form.startPage.colorTheme}
+        submissionData={submissionData}
+        formTitle={form.title}
+        endPage={form.endPage}
+        isFeedbackSubmitted={isFeedbackSubmitted}
+        handleSubmitFeedback={handleSubmitFeedback}
+      />
+    </Box>
+  )
 }
