+++ frontend/src/features/public-form/components/FormEndPage/components/PaymentsThankYouSvgr.tsx
@@ -1,0 +1,21 @@
+import { SVGProps } from 'react'
+import { chakra } from '@chakra-ui/react'
+
+export const PaymentsThankYouSvgr = chakra((props: SVGProps<SVGSVGElement>) => (
+  <svg
+    width="64"
+    height="64"
+    viewBox="0 0 64 64"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <g id="Icon/check-circle-solid">
+      <path
+        id="Vector"
+        d="M32.0006 5.33203C17.2967 5.33203 5.33398 17.2947 5.33398 31.9987C5.33398 46.7027 17.2967 58.6654 32.0006 58.6654C46.7046 58.6654 58.6673 46.7027 58.6673 31.9987C58.6673 17.2947 46.7046 5.33203 32.0006 5.33203ZM26.67 43.7667L16.7687 33.8867L20.534 30.1107L26.6646 36.2307L40.782 22.1134L44.5527 25.884L26.67 43.7667Z"
+        fill="#05CC9A"
+      />
+    </g>
+  </svg>
+))
